api/requirements.txt
fastapi==0.115.12
pdfplumber==0.11.6
pydantic==2.11.5
python-dotenv==1.1.0
qdrant_client==1.14.2
sentence_transformers==4.1.0
spacy==3.8.7
together==1.5.8
torch==2.6.0
transformers==4.52.4
uvicorn[standard]
fastapi[standard]
google-genai==1.30.0
